Experimenting with separating the printing and counting.
At the moment there is no printing done, but this looks promising: no
printing means that we don't have to keep the current path around (the
visited array is sufficient), fewer conditional checks, no arithmetic to
keep track of the number of paths we've already printed, etc.

Next step is to add a separate routine for doing the printing. The step
after that is to choose between separate and in-line printing based on
how many paths we have to print out.