Added multithreading support.
There's a command line option, -t, to use the multithreaded counting. This is significantly SLOWER than the single threaded version at the moment, by a whole order of magnitude.

The visited array is extremely efficient in the single-threaded case. For the multithreaded case we either have to copy it each time we split off a new thread/task/whatever, or use something else. The former is extremely slow. The best "something else" I've been able to come up with so far is the DFSTree structure (still really slow).

The main reason the DFSTree is slow is that you have to walk back up the tree to see whether you've already visited a node, instead of it being a direct array access. Optimising this is very important for getting decent performance. Maybe we could store a list of active path IDs for each node, instead of a single flag?

The granularity of the tasks at the moment is way too fine. It should be processing at least two or three levels of the tree in each task, instead of splitting off a task for every node (depending on the cardinality of the nodes).

I think the way to get good threaded performance for this problem is to only split up the work at the very topmost levels of the tree. Give each core its own visited array and effectively just run as many instances of the serial code as there are processors.

We'll see.