Threading strategy that actually works!
Finally came up with a threading strategy that actually makes it go
faster instead of slower! :-)

The idea is to do breadth-first search of the graph, adding the nodes we
find to a queue until the queue gets big enough to be worth processing
in parallel. This is done by adding the root node to begin with, then
replacing the node from the front of the queue and pushing each of its
children, until the queue gets big enough or it empties out.

Once it's big enough, we do a tbb::parallel_reduce over the nodes in the
queue. Each of the functor objects has its own visited array. The join
step is just adding the count from the RHS to the count of the LHS.

It's not perfect yet though: it crashes on the 20-30 test data. Need to
fix that and then look at speeding up the traversal for printing.