Experimenting with alternative types as an optimisation.
Added a FastVector class, with a similar interface to a std::vector. This
stores the data as an array plus a length, whereas the STL version stores it
as start and end pointers. This should mean that taking the size of the array
is a bit more efficient - it's a single read, rather than two reads and a
subtraction.

Added a typedef for the visited arrays, so I could try out different types for
them easily. It looks like bool is the fastest, with uint64_t coming second.

Also fixed a deletion bug in the CountPathsFunctor's destructor it was doing a
pointer-style delete instead of an array-style delete.